agent-service/requirements.txt
Flask
Flask-cors
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai 
langchain-groq